2019/python-auth-samples/requirements.txt
google-auth==1.6.3
pydata-google-auth==0.1.2
requests==2.21.0